0. ToOne [#6873]

+++ src/main/java/ru/job4j/todo/controller/TaskController.java
@@ -59,7 +59,11 @@ public class TaskController {
     }
 
     @PostMapping("/task/update")
-    public String update(@ModelAttribute Task task, Model model, @SessionAttribute User user) {
+    public String update(@ModelAttribute Task task, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            return "redirect:/users/login";
+        }
         model.addAttribute("user", user);
         task.setUser(user);
         var isUpdated = taskService.update(task);
@@ -71,7 +75,11 @@ public class TaskController {
     }
 
     @GetMapping("/task/delete/{id}")
-    public String delete(Model model, @PathVariable int id) {
+    public String delete(Model model, @PathVariable int id, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            return "redirect:/users/login";
+        }
         var isDeleted = taskService.deleteById(id);
         if (!isDeleted) {
             model.addAttribute("message", "Failed to delete the task");
@@ -81,7 +89,11 @@ public class TaskController {
     }
 
     @GetMapping("/task/complete/{id}")
-    public String completeTask(Model model, @ModelAttribute Task task) {
+    public String completeTask(Model model, @ModelAttribute Task task, HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user == null) {
+            return "redirect:/users/login";
+        }
         var isUpdated = taskService.completeTask(task);
         if (!isUpdated) {
             model.addAttribute("message", "Failed to complete the task");
